requirements.txt
fastapi
uvicorn[standard]
qdrant-client
torch
sentence-transformers
python-multipart
python-dotenv